rename: 프로젝트 패키지 분리 및 이동
